Allow any template for project assistant sessions via shared template picker
